Refactor ParsimonyModel trait to diff file
